==> step_1_pipeline.ipynb <==
# Курсовой проект по циклу "Машинное обучение в бизнесе"

## Прогнозирование рака легких 

датасет: https://www.kaggle.com/datasets/nancyalaswad90/lung-cancer

legend:

1. Gender: M(male), F(female) 
2. Age: Age of the patient 
3. Smoking: YES=2 , NO=1. 
4. Yellow fingers: YES=2 , NO=1. 
5. Anxiety: YES=2 , NO=1. 
6. Peer_pressure: YES=2 , NO=1. 
7. Chronic Disease: YES=2 , NO=1. 
8. Fatigue: YES=2 , NO=1. 
9. Allergy: YES=2 , NO=1. 
10. Wheezing: YES=2 , NO=1. 
11. Alcohol: YES=2 , NO=1. 
12. Coughing: YES=2 , NO=1. 
13. Shortness of Breath: YES=2 , NO=1. 
14. Swallowing Difficulty: YES=2 , NO=1. 
15. Chest pain: YES=2 , NO=1. 
16. Lung Cancer: YES , NO. - tareget

имортируем необходимые бибилотеки и модули

import pandas as pd
import numpy as np

import dill


from sklearn.model_selection import train_test_split
from sklearn.pipeline import Pipeline, FeatureUnion
from sklearn.base import BaseEstimator, TransformerMixin
from sklearn.model_selection import cross_val_score, train_test_split, GridSearchCV 

# from sklearn.ensemble import RandomForestClassifier

from sklearn.ensemble import GradientBoostingClassifier

from sklearn.metrics import f1_score, roc_auc_score, precision_score, classification_report, precision_recall_curve, confusion_matrix, log_loss, roc_curve

import itertools

import seaborn as sns
import matplotlib.pyplot as plt

import warnings
warnings.filterwarnings("ignore")

dataset = pd.read_csv('survey_lung_cancer.csv')
dataset.head(5)

dataset.info()

проверим наличие пропусков

dataset.isnull().sum()

соотношение целевой перменной

dataset['LUNG_CANCER'].value_counts()

numeric_cols=dataset.select_dtypes(include=np.number).columns.tolist()
categorial_cols=dataset.select_dtypes(object).columns.tolist()

print(f'Вывод: \nДатасет содержит 309 строк (наблюдений), 16 столбцов, в том числе: 15 признаков и 1 - целевая переменная.\n\n' 
      f'Тип данных int64 встречается у {len(numeric_cols)} столбцов: {numeric_cols},\n\nobject - у {len(categorial_cols)}: {categorial_cols}.\n\n'
      f'Пропуски в датасете отсутствуют.\n\n'
     f'Соотношение классов целевой переменной не сбалансировано')

## EDA (exploratory data analysis)

Категориальные признаки

print(f'{len(categorial_cols)} категориальных признака\n{categorial_cols}\n')

for i in categorial_cols:
    unique_no = dataset[i].nunique()
    unique_name = dataset[i].unique().tolist()
    print(f'{i}, имеет {unique_no} уникальных значения:\n{unique_name}\n')

## Feature engineering

переименуем столбцы

dataset.columns = ['GENDER', 'AGE', 'SMOKING','YELLOW_FINGERS', 'ANXIETY', 'PEER_PRESSURE', 'CHRONIC_DISEASE', 'FATIGUE',
                   'ALLERGY', 'WHEEZING', 'ALCOHOL_CONSUMING', 'COUGHING', 'SHORTNESS_OF_BREATH', 
                   'SWALLOWING_DIFFICULTY', 'CHEST_PAIN', 'LUNG_CANCER']

изменим значения полей, содержащих категориальные признаки на бинарные (исходные: YES = 2 и NO = 1 на YES = 1 и NO = 0)

поля: 'SMOKING','YELLOW_FINGERS', 'ANXIETY', 'PEER_PRESSURE', 'CHRONIC DISEASE', 'FATIGUE', 'ALLERGY',
'WHEEZING', 'ALCOHOL CONSUMING', 'COUGHING', 'SHORTNESS OF BREATH', 'SWALLOWING DIFFICULTY', 'CHEST PAIN' 

list_feat_names = ['SMOKING','YELLOW_FINGERS', 'ANXIETY', 'PEER_PRESSURE', 'CHRONIC_DISEASE', 'FATIGUE',
                   'ALLERGY', 'WHEEZING', 'ALCOHOL_CONSUMING', 'COUGHING', 'SHORTNESS_OF_BREATH', 
                   'SWALLOWING_DIFFICULTY', 'CHEST_PAIN']


for name in list_feat_names:
    dataset[name] = dataset[name].replace([2, 1],[1, 0])

dataset['GENDER'] = dataset['GENDER'].map({'M': 1, 'F':0})

dataset['LUNG_CANCER'] = dataset['LUNG_CANCER'].map({'YES': 1, 'NO':0})

row_40 = dataset[dataset['AGE'] <= 40 ].index
dataset = dataset.drop(dataset.index[row_40])

row_80 = dataset[dataset['AGE'] > 80].index
dataset = dataset.drop(dataset.index[row_80])

dataset.head(3)

dataset.info()

разделим датасет на тренировочную и тестовую выборку 

features = ['GENDER', 'AGE', 'SMOKING','YELLOW_FINGERS', 'ANXIETY', 'PEER_PRESSURE', 'CHRONIC_DISEASE', 'FATIGUE',
                   'ALLERGY', 'WHEEZING', 'ALCOHOL_CONSUMING', 'COUGHING', 'SHORTNESS_OF_BREATH', 
                   'SWALLOWING_DIFFICULTY', 'CHEST_PAIN']
target = 'LUNG_CANCER'


X_train, X_test, y_train, y_test = train_test_split(dataset[features], dataset[target], test_size=0.33, random_state=42,
                                                   stratify=dataset[target])

#save test
X_test.to_csv("X_test.csv", index=None)
y_test.to_csv("y_test.csv", index=None)
#save train
X_train.to_csv("X_train.csv", index=None)
y_train.to_csv("y_train.csv", index=None)

class ColumnSelector(BaseEstimator, TransformerMixin):
    """
    Transformer to select a single column from the data frame to perform additional transformations on
    """
    def __init__(self, key):
        self.key = key

    def fit(self, X, y=None):
        return self

    def transform(self, X):
        return X[self.key]
    
class NumberSelector(BaseEstimator, TransformerMixin):
    """
    Transformer to select a single column from the data frame to perform additional transformations on
    Use on numeric columns in the data
    """
    def __init__(self, key):
        self.key = key

    def fit(self, X, y=None):
        return self

    def transform(self, X):
        return X[[self.key]]
    
class OHEEncoder(BaseEstimator, TransformerMixin):
    def __init__(self, key):
        self.key = key
        self.columns = []

    def fit(self, X, y=None):
        self.columns = [col for col in pd.get_dummies(X, prefix=self.key).columns]
        return self

    def transform(self, X):
        X = pd.get_dummies(X, prefix=self.key)
        test_columns = [col for col in X.columns]
        for col_ in test_columns:
            if col_ not in self.columns:
                X[col_] = 0
        return X[self.columns]

gender = Pipeline([
                ('selector', ColumnSelector(key='GENDER')),
                ('ohe', OHEEncoder(key='GENDER'))
            ])

gender.fit_transform(X_train)

number_features = Pipeline([
                ('selector', ColumnSelector(key=['AGE', 'SMOKING','YELLOW_FINGERS', 
                                                 'ANXIETY', 'PEER_PRESSURE', 
                                                 'CHRONIC_DISEASE', 'FATIGUE',
                                                 'ALLERGY', 'WHEEZING', 
                                                 'ALCOHOL_CONSUMING', 
                                                 'COUGHING', 'SHORTNESS_OF_BREATH',
                                                 'SWALLOWING_DIFFICULTY', 'CHEST_PAIN']))])
feats = FeatureUnion([('continuos_features', number_features),
                      ('gender', gender)]) 
feature_processing = Pipeline([('feats', feats)])
feature_processing.fit_transform(X_train)

%%time

pipeline = Pipeline([
    ('features', feats),
#     ('classifier', RandomForestClassifier(random_state = 42)),
    ('classifier', GradientBoostingClassifier(random_state = 42)),    
])

#запустим кросс-валидацию
cv_scores = cross_val_score(pipeline, X_train, y_train, cv=16, scoring='roc_auc')
cv_score = np.mean(cv_scores)
cv_score_std = np.std(cv_scores)
print('CV score is {}+-{}'.format(cv_score, cv_score_std))

#обучим пайплайн на всем тренировочном датасете
pipeline.fit(X_train, y_train)
y_score = pipeline.predict_proba(X_test)[:, 1]



Посмотрим, как выглядит наш pipeline

pipeline.steps

Сохраним модель (пайплайн)

with open("GBC_model.dill", "wb") as f:
    dill.dump(pipeline, f)
