==> step_2_predict.ipynb <==
# Курсовой проект по циклу "Машинное обучение в бизнесе"

## Прогнозирование рака легких 

import pandas as pd

import requests

import urllib.request
import json 

dataset = pd.read_csv('X_train.csv')
dataset.head(5)

data = (0, 45, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1, 0, 1)

# формируем запрос
def send_json(x):
    GENDER, AGE, SMOKING, YELLOW_FINGERS, ANXIETY, PEER_PRESSURE, CHRONIC_DISEASE, FATIGUE, ALLERGY, 
    WHEEZING, ALCOHOL_CONSUMING, COUGHING, SHORTNESS_OF_BREATH, SWALLOWING_DIFFICULTY, CHEST_PAIN = x 
    
    print(GENDER, AGE, SMOKING, YELLOW_FINGERS, ANXIETY, PEER_PRESSURE, CHRONIC_DISEASE, FATIGUE, ALLERGY, 
    WHEEZING, ALCOHOL_CONSUMING, COUGHING, SHORTNESS_OF_BREATH, SWALLOWING_DIFFICULTY, CHEST_PAIN)
    
    body = {
        'GENDER' : GENDER,
         'AGE' : AGE,
         'SMOKING' : SMOKING,
         'YELLOW_FINGERS' : YELLOW_FINGERS,
         'ANXIETY' : ANXIETY,
         'PEER_PRESSURE' : PEER_PRESSURE,
         'CHRONIC_DISEASE' : CHRONIC_DISEASE,
         'FATIGUE' : FATIGUE,
         'ALLERGY' : ALLERGY,
         'WHEEZING' : WHEEZING,
         'ALCOHOL_CONSUMING' : ALCOHOL_CONSUMING,
         'COUGHING' : COUGHING,
         'SHORTNESS_OF_BREATH' : SHORTNESS_OF_BREATH,
         'SWALLOWING_DIFFICULTY' : SWALLOWING_DIFFICULTY,
         'CHEST_PAIN' : CHEST_PAIN
        }
    
    headers = {'content-type': 'application/json; charset=utf-8'}
    response = requests.post(myurl, json=body, headers=headers)
    return response.json()['predictions']

response = send_json(data)
print('предсказание', response)

import numpy as np

precision, recall, thresholds = precision_recall_curve(y_test[:N], predictions)

fscore = (2 * precision * recall) / (precision + recall)
# locate the index of the largest f score
ix = np.argmax(fscore)
print(f'Best Threshold={thresholds[ix]}, F-Score={fscore[ix]:.3f}, Precision={precision[ix]:.3f}, Recall={recall[ix]:.3f}')

roc_auc_score(y_score=predictions.values, y_true=y_test.values[:N])
