==> step_3_server.ipynb <==
# Курсовой проект по циклу "Машинное обучение в бизнесе"

## Прогнозирование рака легких 

#from flask_ngrok import run_with_ngrok
from flask import Flask, request, jsonify
import pandas as pd

import dill

import warnings
warnings.filterwarnings("ignore")

app = Flask(__name__)

@app.route("/a")
def hello():
    return "Hello World!"

if __name__ == '__main__':
    app.run()

X_test = pd.read_csv("X_test.csv")
y_test = pd.read_csv("y_test.csv")

app = Flask(__name__)


@app.route("/", methods=["GET"])
def general():
    return "Welcome to prediction process"

@app.route('/predict', methods=['POST'])
def predict():
    data = {"success": False}

    GENDER, AGE, SMOKING, YELLOW_FINGERS, ANXIETY, 
    PEER_PRESSURE, CHRONIC_DISEASE, FATIGUE, ALLERGY, WHEEZING, 
    ALCOHOL_CONSUMING, COUGHING, SHORTNESS_OF_BREATH, 
    SWALLOWING_DIFFICULTY, CHEST_PAIN = "", "", "", "", "", "", "", "", "", "", "", "", "", "", ""
    
    request_json = request.get_json()
    
    if request_json["GENDER"]:
        GENDER = request_json['GENDER']
    
    if request_json["AGE"]:
        AGE = request_json['AGE']
                
    if request_json["SMOKING"]:
        SMOKING = request_json['SMOKING']

    if request_json["YELLOW_FINGERS"]:
        YELLOW_FINGERS = request_json['YELLOW_FINGERS']
    
    if request_json["ANXIETY"]:
        ANXIETY = request_json['ANXIETY']
                
    if request_json["PEER_PRESSURE"]:
        PEER_PRESSURE = request_json['PEER_PRESSURE']   
    
    if request_json["CHRONIC_DISEASE"]:
        CHRONIC_DISEASE = request_json['CHRONIC_DISEASE']
    
    if request_json["FATIGUE"]:
        FATIGUE = request_json['FATIGUE']
                
    if request_json["ALLERGY"]:
        ALLERGY = request_json['ALLERGY']
        
    if request_json["WHEEZING"]:
        WHEEZING = request_json['WHEEZING']
    
    if request_json["ALCOHOL_CONSUMING"]:
        ALCOHOL_CONSUMING = request_json['ALCOHOL_CONSUMING']
                
    if request_json["COUGHING"]:
        COUGHING = request_json['COUGHING']
        
    if request_json["SHORTNESS_OF_BREATH"]:
        SHORTNESS_OF_BREATH = request_json['SHORTNESS_OF_BREATH']
    
    if request_json["SWALLOWING_DIFFICULTY"]:
        SWALLOWING_DIFFICULTY = request_json['SWALLOWING_DIFFICULTY']
                
    if request_json["CHEST_PAIN"]:
        CHEST_PAIN = request_json['CHEST_PAIN']
        
    print(description)  
    preds = model.predict_proba(pd.DataFrame({'GENDER': [GENDER],
                                             'AGE': [AGE],
                                             'SMOKING': [SMOKING],
                                             'YELLOW_FINGERS': [YELLOW_FINGERS],
                                             'ANXIETY': [ANXIETY],
                                             'PEER_PRESSURE': [PEER_PRESSURE],
                                             'CHRONIC_DISEASE': [CHRONIC_DISEASE],
                                             'FATIGUE': [FATIGUE],
                                             'ALLERGY': [ALLERGY],
                                             'WHEEZING': [WHEEZING],
                                             'ALCOHOL_CONSUMING': [ALCOHOL_CONSUMING],
                                             'COUGHING': [COUGHING],
                                             'SHORTNESS_OF_BREATH': [SHORTNESS_OF_BREATH],
                                             'SWALLOWING_DIFFICULTY': [SWALLOWING_DIFFICULTY],
                                             'CHEST_PAIN': [CHEST_PAIN]}))
        

    data["predictions"] = preds[:, 1][0]
    data["description"] = description
    data["success"] = True
    print('OK')

    return jsonify(data)


if __name__ == '__main__':
    app.run()
